requirements.txt
pandas>=1.5.0
numpy>=1.22.0
matplotlib>=3.5.0
requests>=2.28.0
python-dotenv>=0.20.0
paho-mqtt>=2.0.0
